CollectionView可顯示內容

+++ Coffee_order_system/Model/Struct.swift
@@ -12,7 +12,7 @@ import Foundation
 struct Items:Codable {
     
     var image:String
-    var price:Double
+    var price:Int
     var name:String
     
 }
